+++ src/KardFinancial.Test/Unit/MockServer/Organizations/Placements/ListTest.cs
@@ -22,6 +22,7 @@ public class ListTest : BaseMockServerTest
                     "name": "name",
                     "organizationId": "organizationId",
                     "availableSlots": 1,
+                    "contentStrategyId": "contentStrategyId",
                     "createdAt": "2024-01-15T09:30:00.000Z",
                     "lastModified": "2024-01-15T09:30:00.000Z"
                   }
@@ -33,6 +34,7 @@ public class ListTest : BaseMockServerTest
                     "name": "name",
                     "organizationId": "organizationId",
                     "availableSlots": 1,
+                    "contentStrategyId": "contentStrategyId",
                     "createdAt": "2024-01-15T09:30:00.000Z",
                     "lastModified": "2024-01-15T09:30:00.000Z"
                   }