+++ src/KardFinancial.Test/Unit/MockServer/Organizations/Placements/UpdateTest.cs
@@ -32,6 +32,7 @@ public class UpdateTest : BaseMockServerTest
                 "name": "name",
                 "organizationId": "organizationId",
                 "availableSlots": 1,
+                "contentStrategyId": "contentStrategyId",
                 "createdAt": "2024-01-15T09:30:00.000Z",
                 "lastModified": "2024-01-15T09:30:00.000Z"
               }